+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org).
 
+[v0.2.0]
+--------
+
+### Added
+- `$event->handleValue($value)` shortcut for setting the `value`, and setting
+  `$event->handled = true`.
+
 
 [v0.1.0]
 --------